+++ neonrpg/Block/Blocks/BlockBrick.cs
@@ -7,6 +7,6 @@ namespace neonrpg.Block.Blocks {
 
     class BlockBrick : BaseBlock {
 
-        public BlockBrick(uint x, uint y) : base(2, "Brick", x, y, '#', Color.WHITE, new Color("230", "70", "70"), canWalkThrough: false) { }
+        public BlockBrick(ushort x, ushort y) : base(2, "Brick", x, y, '#', Color.WHITE, new Color("230", "70", "70"), canWalkThrough: false) { }
     }
 }
