+++ neonrpg/Block/Blocks/BlockFlower.cs
@@ -7,7 +7,7 @@ namespace neonrpg.Block.Blocks {
 
     class BlockFlower : BaseBlock {
 
-        public BlockFlower(uint x, uint y, byte data) : base(1, "Flower", x, y, '*', Color.YELLOW, new Color("0", "140", "10"), data: data) {
+        public BlockFlower(ushort x, ushort y, byte data) : base(1, "Flower", x, y, '*', Color.YELLOW, new Color("0", "140", "10"), data: data) {
             switch(data) {
                 case 0:
                     Foreground = Color.YELLOW;
