+++ neonrpg/Block/Blocks/BlockGrass.cs
@@ -7,6 +7,6 @@ namespace neonrpg.Block.Blocks {
 
     class BlockGrass : BaseBlock {
 
-        public BlockGrass(uint x, uint y) : base(0, "Grass", x, y, ' ', new Color("0", "140", "10"), new Color("0", "140", "10")) { }
+        public BlockGrass(ushort x, ushort y) : base(0, "Grass", x, y, ' ', new Color("0", "140", "10"), new Color("0", "140", "10")) { }
     }
 }
